refact routes, more friendly to forecast and truth values
