+++ src/tests/test_main.py
@@ -14,59 +14,3 @@ def test_read_main():
     response = client.get("/")
     assert response.status_code == 200
     assert response.json()["version"] == version
-
-
-def test_read_latest_one_gsp(db_session):
-    """Check main GB/pv/gsp/{gsp_id} route works"""
-
-    forecasts = make_fake_forecasts(gsp_ids=list(range(0, 10)), session=db_session)
-    db_session.add_all(forecasts)
-
-    app.dependency_overrides[get_session] = lambda: db_session
-
-    response = client.get("/v0/forecasts/GB/pv/gsp/1")
-    assert response.status_code == 200
-
-    _ = Forecast(**response.json())
-
-
-def test_read_latest_all_gsp(db_session):
-    """Check main GB/pv/gsp route works"""
-
-    forecasts = make_fake_forecasts(gsp_ids=list(range(0, 338)), session=db_session)
-    db_session.add_all(forecasts)
-
-    app.dependency_overrides[get_session] = lambda: db_session
-
-    response = client.get("/v0/forecasts/GB/pv/gsp")
-    assert response.status_code == 200
-
-    r = ManyForecasts(**response.json())
-    assert len(r.forecasts) == 338
-
-
-def test_read_latest_national(db_session):
-    """Check main GB/pv/national route works"""
-
-    forecast = make_fake_national_forecast()
-    db_session.add(forecast)
-
-    app.dependency_overrides[get_session] = lambda: db_session
-
-    response = client.get("/v0/forecasts/GB/pv/national")
-    assert response.status_code == 200
-
-    _ = Forecast(**response.json())
-
-
-def test_gsp_boundaries(db_session):
-    """Check main GB/pv/national route works"""
-
-    forecast = make_fake_national_forecast()
-    db_session.add(forecast)
-
-    app.dependency_overrides[get_session] = lambda: db_session
-
-    response = client.get("/v0/forecasts/GB/pv/gsp_boundaries")
-    assert response.status_code == 200
-    assert len(response.json()) > 0
